Move global maps to static locals in free functions
That allows to avoid initialization races on usages.
