+++ sycl/source/detail/device_filter.cpp
@@ -30,11 +30,11 @@ device_filter::device_filter(const std::string &FilterString) {
 
   // Handle the optional 1st field of the filter, backend
   // Check if the first entry matches with a known backend type
-  auto It =
-      std::find_if(std::begin(SyclBeMap), std::end(SyclBeMap), findElement);
+  auto It = std::find_if(std::begin(getSyclBeMap()), std::end(getSyclBeMap()),
+                         findElement);
   // If no match is found, set the backend type backend::all
   // which actually means 'any backend' will be a match.
-  if (It == SyclBeMap.end())
+  if (It == getSyclBeMap().end())
     Backend = backend::all;
   else {
     Backend = It->second;
@@ -49,11 +49,11 @@ device_filter::device_filter(const std::string &FilterString) {
   if (Cursor >= FilterString.size()) {
     DeviceType = info::device_type::all;
   } else {
-    auto Iter = std::find_if(std::begin(SyclDeviceTypeMap),
-                             std::end(SyclDeviceTypeMap), findElement);
+    auto Iter = std::find_if(std::begin(getSyclDeviceTypeMap()),
+                             std::end(getSyclDeviceTypeMap()), findElement);
     // If no match is found, set device_type 'all',
     // which actually means 'any device_type' will be a match.
-    if (Iter == SyclDeviceTypeMap.end())
+    if (Iter == getSyclDeviceTypeMap().end())
       DeviceType = info::device_type::all;
     else {
       DeviceType = Iter->second;
